moved fetcher package into timeline
