+++ app/src/main/java/com/freshdigitable/udonroad/module/AppComponent.java
@@ -27,10 +27,10 @@ import com.freshdigitable.udonroad.UserInfoFragment;
 import com.freshdigitable.udonroad.UserInfoPagerFragment;
 import com.freshdigitable.udonroad.UserSettingsActivity;
 import com.freshdigitable.udonroad.UserStreamUtil;
-import com.freshdigitable.udonroad.fetcher.ListFetcher;
-import com.freshdigitable.udonroad.fetcher.ListsListFetcherModule;
-import com.freshdigitable.udonroad.fetcher.StatusListFetcherModule;
-import com.freshdigitable.udonroad.fetcher.UserListFetcherModule;
+import com.freshdigitable.udonroad.timeline.fetcher.ListFetcher;
+import com.freshdigitable.udonroad.timeline.fetcher.ListsListFetcherModule;
+import com.freshdigitable.udonroad.timeline.fetcher.StatusListFetcherModule;
+import com.freshdigitable.udonroad.timeline.fetcher.UserListFetcherModule;
 import com.freshdigitable.udonroad.input.TweetInputActivity;
 import com.freshdigitable.udonroad.input.TweetInputFragment;
 import com.freshdigitable.udonroad.media.MediaViewActivity;
