+++ app/src/main/java/com/freshdigitable/udonroad/subscriber/ListsListRequestWorker.java
@@ -21,8 +21,8 @@ import android.support.annotation.NonNull;
 import com.freshdigitable.udonroad.R;
 import com.freshdigitable.udonroad.StoreType;
 import com.freshdigitable.udonroad.datastore.WritableSortedCache;
-import com.freshdigitable.udonroad.fetcher.FetchQuery;
-import com.freshdigitable.udonroad.fetcher.ListFetcher;
+import com.freshdigitable.udonroad.timeline.fetcher.FetchQuery;
+import com.freshdigitable.udonroad.timeline.fetcher.ListFetcher;
 import com.freshdigitable.udonroad.ffab.IndicatableFFAB.OnIffabItemSelectedListener;
 
 import java.util.Collection;
