+++ app/src/main/java/com/freshdigitable/udonroad/timeline/fetcher/ListFetcherModuleKey.java
@@ -14,17 +14,25 @@
  * limitations under the License.
  */
 
-package com.freshdigitable.udonroad.fetcher;
+package com.freshdigitable.udonroad.timeline.fetcher;
 
-import java.util.List;
+import com.freshdigitable.udonroad.StoreType;
 
-import io.reactivex.Single;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import dagger.MapKey;
 
 /**
  * Created by akihit on 2018/01/10.
  */
-public interface ListFetcher<T> {
-  Single<? extends List<T>> fetchInit(FetchQuery query);
-
-  Single<? extends List<T>> fetchNext(FetchQuery query);
+@Documented
+@Target({ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+@MapKey
+@interface ListFetcherModuleKey {
+  StoreType value();
 }
